Combine packets into one file
